継続日数のロジック分離
